conditional navbar working with initial landing pages for each link
