+++ website/templates/cart.html
@@ -1,0 +1,6 @@
+{% extends "index.html" %}
+
+{% block content %}
+<p> {{cart}}</p>
+
+{% endblock content %}
