+++ website/templates/categories.html
@@ -1,0 +1,6 @@
+{% extends "index.html" %}
+
+{% block content %}
+<p> {{categories}}</p>
+
+{% endblock content %}
