+++ website/templates/nav.html
@@ -1,0 +1,50 @@
+<!-- Basic block for navbar -->
+{% block nav %}
+{% if user.is_authenticated %}
+<nav class="=navbar navbar-expand-lg navbar-light bg-light nav_flex">
+
+    <ul class="navbar-nav nav_flex">
+        <li>
+            <a class="navbar-brand" href="{% url 'website:index' %}">Bangazon</a>
+        </li>
+        <li class="nav-item">
+            <a class="nav-link" href="{% url 'website:cart' user.id %}">Cart</a>
+        </li>
+        <li class="nav-item">
+            <a class="nav-link" href="{% url 'website:sell' %}">Sell</a>
+        </li>
+        <li class="nav-item">
+            <a class="nav-link" href="{% url 'website:categories' %}">Categories</a>
+        </li>
+        <li class="nav-item">
+            <a class="nav-link" href="{% url 'website:profile' user.id %}">Profile</a>
+        </li>
+        <li class="nav-item">
+            <a class="nav-link" href="{% url 'website:logout' %}">Logout</a>
+        </li>
+    </ul>
+    </div>
+</nav>
+
+{% else %}
+<nav class="=navbar navbar-expand-lg navbar-light bg-light nav_flex">
+
+    <ul class="navbar-nav nav_flex">
+        <li>
+            <a class="navbar-brand" href="{% url 'website:index' %}">Bangazon</a>
+        </li>
+        <li class="nav-item">
+            <a class="nav-link" href="#">Categories</a>
+        </li>       
+        <li class="nav-item">
+            <a class="nav-link" href="{% url 'website:login' %}">Login</a>
+        </li>
+        <li class="nav-item">
+            <a class="nav-link" href="{% url 'website:register' %}">Register</a>
+        </li>
+    </ul>
+    </div>
+</nav>
+{% endif %}
+
+{% endblock %}
