+++ website/templates/profile.html
@@ -1,0 +1,6 @@
+{% extends "index.html" %}
+
+{% block content %}
+<p> {{profile}}</p>
+
+{% endblock content %}
