+++ website/urls.py
@@ -5,9 +5,13 @@ from . import views
 app_name = "website"
 urlpatterns = [
     url(r'^$', views.index, name='index'),
+
     url(r'^login$', views.login_user, name='login'),
     url(r'^logout$', views.user_logout, name='logout'),
     url(r'^register$', views.register, name='register'),
     url(r'^sell$', views.sell_product, name='sell'),
     url(r'^products$', views.list_products, name='list_products'),
+    url(r'^cart/(?P<pk>\d+)/$', views.cart, name="cart"),
+    url(r'^categories$', views.categories, name="categories"),
+    url(r'^profile/(?P<pk>\d+)/$', views.profile, name="profile")
 ]
