+++ website/views/__init__.py
@@ -1,0 +1,6 @@
+from .cart_view import *
+from .handle_user_views import *
+from .main_views import *
+from .products_views import *
+from .profile_views import *
+from .sell_views import *
