+++ website/views/products_views.py
@@ -1,0 +1,18 @@
+from django.contrib.auth import logout, login, authenticate
+from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse, HttpResponseRedirect, Http404
+from django.shortcuts import render
+from django.template import RequestContext
+
+from website.forms import UserForm, ProductForm
+from website.models import Product
+
+
+def list_products(request):
+    all_products = Product.objects.all()
+    template_name = 'product/list.html'
+    return render(request, template_name, {'products': all_products})
+
+def categories(request):
+    context = {"categories":"all the things"}
+    return render(request, 'categories.html', context)
