+++ website/views/profile_views.py
@@ -1,0 +1,15 @@
+from django.contrib.auth import logout, login, authenticate
+from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse, HttpResponseRedirect, Http404
+from django.shortcuts import render
+from django.template import RequestContext
+
+from website.forms import UserForm, ProductForm
+from website.models import Product
+
+def profile(request, pk):
+    context = {"profile":"its me"}
+    return render(request, 'profile.html', context)
+
+  
+
